Only require solidus_core
Requiring solidus istead of `solidus_core` requires a ton of code that
might not even be used in the application requiring this extension.

By requiring just `solidus_core` life will be easier for people that
are, for example, not using solidus_frontend.

The new `solidus_dev_support` gem has been used in order to make the
specs pass.